+++ source/qdk_package/qdk/openqasm/_native_syntax.py
@@ -1,0 +1,111 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+# Re-exports the flat OpenQASM native surface so its type declarations can live
+# in the sibling `_native_syntax.pyi` stub instead of the shared `_native.pyi`.
+
+# pyright: reportAttributeAccessIssue=false
+
+from .._native import (
+    Span,
+    PositionEncoding,
+    ClassicalType,
+    AccessControl,
+    IntType,
+    UintType,
+    FloatType,
+    AngleType,
+    BitType,
+    ComplexType,
+    BoolType,
+    DurationType,
+    StretchType,
+    QubitType,
+    ErrorType,
+    ArrayType,
+    StaticArrayReferenceType,
+    DynArrayReferenceType,
+    ResolutionStatus,
+    Position,
+    SourceRange,
+    SourceFile,
+    SourceMap,
+    SourceDocument,
+    Severity,
+    Label,
+    Diagnostic,
+    QASMNode,
+    Expression,
+    Statement,
+    Annotation,
+    Program,
+    QuantumGateModifier,
+    RangeDefinition,
+    DiscreteSet,
+    IndexList,
+    SwitchCase,
+    SubroutineParameter,
+    Identifier,
+    IndexedIdentifier,
+    HardwareQubit,
+    ErrorExpression,
+    UnaryExpression,
+    BinaryExpression,
+    BinaryOperator,
+    UnaryOperator,
+    TimeUnit,
+    IOKeyword,
+    GateModifierName,
+    IntegerLiteral,
+    FloatLiteral,
+    ImaginaryLiteral,
+    BooleanLiteral,
+    BitstringLiteral,
+    DurationLiteral,
+    ArrayLiteral,
+    StringLiteral,
+    FunctionCall,
+    Cast,
+    IndexExpression,
+    ParenExpression,
+    DurationOf,
+    Concatenation,
+    QuantumMeasurement,
+    QubitDeclaration,
+    AliasStatement,
+    ClassicalAssignment,
+    CompoundAssignment,
+    QuantumBarrier,
+    Box,
+    BreakStatement,
+    CompoundStatement,
+    CalibrationStatement,
+    CalibrationGrammarDeclaration,
+    ClassicalDeclaration,
+    ConstantDeclaration,
+    ContinueStatement,
+    SubroutineDefinition,
+    CalibrationDefinition,
+    DelayInstruction,
+    EndStatement,
+    ExpressionStatement,
+    ExternDeclaration,
+    ForInLoop,
+    BranchingStatement,
+    QuantumGate,
+    QuantumPhase,
+    Include,
+    IODeclaration,
+    QuantumMeasurementStatement,
+    Pragma,
+    QuantumGateDefinition,
+    QuantumReset,
+    ReturnStatement,
+    SwitchStatement,
+    WhileLoop,
+    ErrorStatement,
+    ParseResult,
+    _QASMUnparseError,
+    parse,
+    qasm_dumps,
+)
