+++ source/qdk_package/tests/test_qasm_semantic_visitors.py
@@ -1,166 +1,18 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 
-"""Proves that every ``visit_<NodeType>`` callback of the semantic visitor is
-dispatched correctly.
+"""Guards how the semantic visitor dispatches and traverses.
 
-A comprehensive `OpenQASM` program is analyzed, and a :class:`QASMVisitor`
-subclass is built with a ``visit_<NodeType>`` method for *every* semantic node
-class exported by :mod:`qdk.openqasm.semantic`. Each callback records that it
-fired and asserts it received a node of exactly its own type. The test then
-checks that the set of callbacks that fired equals the set of node types
-actually present in the tree (collected independently), and that both counts
-match.
+The analyzed tree differs from the parsed one in ways a visitor must handle:
+broadcast gate calls expand into one node per qubit, and declared widths move
+onto the resolved type instead of remaining child expressions. These tests pin
+typed dispatch, context propagation, and single-visit traversal over that tree.
 """
 
-from typing import Any, Callable, Dict, List, Tuple
+from typing import Any, List, Tuple
 
 from qdk.openqasm import semantic
-from qdk.openqasm.semantic import QASMNode, QASMVisitor, Statement
-
-# A single program exercising a broad range of OpenQASM 3 constructs so that as
-# many distinct semantic node types as possible appear in the analyzed tree.
-_PROGRAM = """OPENQASM 3.0;
-include "stdgates.inc";
-extern myext(int) -> int;
-gate mygate(theta) q { rz(theta) q; }
-def add(int a, int b) -> int { return a + b; }
-const int N = 3;
-input int shots;
-output bit result;
-qubit q;
-qubit[2] qs;
-bit c;
-bit[2] cs;
-int a = 1 + 2;
-float f = -1.5;
-int b = (a * 2) - 1;
-int d = add(a, N);
-bit[4] joined = cs ++ cs;
-bool cond = a > 0;
-a = 5;
-a += 1;
-let myalias = qs[0:1];
-x q;
-x $0;
-mygate(0.5) q;
-ctrl @ x qs[0], qs[1];
-h qs[0];
-gphase(pi / 2);
-c = measure q;
-measure qs[0];
-reset q;
-barrier q, qs;
-delay[10ns] q;
-box { x q; }
-pragma qdk.example some custom pragma text
-duration dur = durationof({x q;});
-if (a == 3) { x q; } else { y q; }
-for int i in [0:2] { x q; if (cond) { break; } else { continue; } }
-while (a > 0) { a -= 1; }
-switch (a) { case 1 { x q; } default { y q; } }
-int g = int(f);
-add(a, N);
-end;
-"""
-
-# Core node types the program is guaranteed to produce. Kept as a stable subset
-# so the test remains a meaningful proof of dispatch even if the analyzer gains
-# or loses more exotic node kinds over time.
-_REQUIRED_TYPES = {
-    "Program",
-    "QubitDeclaration",
-    "QubitArrayDeclaration",
-    "ClassicalDeclaration",
-    "QuantumGate",
-    "QuantumGateDefinition",
-    "SubroutineDefinition",
-    "Identifier",
-    "BinaryExpression",
-    "UnaryExpression",
-    "LiteralExpression",
-    "IndexExpression",
-    "ParenExpression",
-    "Pragma",
-    "FunctionCall",
-    "Cast",
-    "BranchingStatement",
-    "ForInLoop",
-    "WhileLoop",
-    "ReturnStatement",
-    "QuantumMeasurement",
-    "QuantumReset",
-    "QuantumBarrier",
-    "ClassicalAssignment",
-    "CompoundStatement",
-}
-
-_MIN_DISTINCT_TYPES = 35
-
-
-def _node_class_names() -> List[str]:
-    """Every semantic node class name exported by ``qdk.openqasm.semantic``."""
-    names = []
-    for name in semantic.__all__:
-        obj = getattr(semantic, name)
-        if isinstance(obj, type) and issubclass(obj, QASMNode):
-            names.append(name)
-    return names
-
-
-def _collect_present_types(program: Any) -> Dict[str, int]:
-    """Independently collects ``type(node).__name__`` -> count over the tree."""
-    counts: Dict[str, int] = {}
-
-    class Collector(QASMVisitor):
-        def generic_visit(self, node: Any) -> None:
-            name = type(node).__name__
-            counts[name] = counts.get(name, 0) + 1
-            super().generic_visit(node)
-
-    Collector().visit(program)
-    return counts
-
-
-def _make_callback(expected: str, dispatched: Dict[str, int], mismatches: List[Tuple[str, str]]) -> Callable:
-    def callback(self: QASMVisitor, node: Any) -> None:
-        actual = type(node).__name__
-        if actual != expected:
-            mismatches.append((expected, actual))
-        dispatched[expected] = dispatched.get(expected, 0) + 1
-        self.generic_visit(node)
-
-    return callback
-
-
-def test_every_semantic_visit_callback_dispatches() -> None:
-    program = semantic.analyze(_PROGRAM).program
-
-    present = _collect_present_types(program)
-    assert _REQUIRED_TYPES <= set(present)
-    assert len(present) >= _MIN_DISTINCT_TYPES
-
-    dispatched: Dict[str, int] = {}
-    mismatches: List[Tuple[str, str]] = []
-
-    # Build a visitor with a visit_<NodeType> callback for every semantic node
-    # class, so no node can fall back to generic_visit unhandled.
-    attrs = {
-        f"visit_{name}": _make_callback(name, dispatched, mismatches)
-        for name in _node_class_names()
-    }
-    all_callbacks_visitor = type("AllSemanticCallbacks", (QASMVisitor,), attrs)
-
-    all_callbacks_visitor().visit(program)
-
-    # Every node was routed to the callback named after its own type.
-    assert mismatches == []
-    # Each present type's callback fired for exactly its nodes, and no present
-    # type was missing a callback.
-    assert dispatched == present
-    # And every present type does have a dedicated callback method.
-    for name in present:
-        assert hasattr(all_callbacks_visitor, f"visit_{name}")
+from qdk.openqasm.semantic import QASMVisitor, Statement
 
 
 def test_broadcast_semantic_visitor_observes_each_projected_gate() -> None:
@@ -285,3 +137,48 @@ int x = 1;
     assert children[0] == declaration.annotations[0]
     assert len(children) > 1
 
+
+def test_auxiliary_node_callbacks_dispatch() -> None:
+    source = """OPENQASM 3.1;
+    include "stdgates.inc";
+    @tag value
+    def f(int[8] a) -> int { return a; }
+    bit[4] bits;
+    let slice = bits[1:2];
+    int selector;
+    switch (selector) { case 1 { selector = 2; } }
+    qubit[2] q;
+    ctrl @ x q[0], q[1];
+    """
+    fired: List[str] = []
+
+    class AuxiliaryVisitor(QASMVisitor):
+        def visit_Annotation(self, node: Any) -> None:
+            fired.append("Annotation")
+
+        def visit_SubroutineParameter(self, node: Any) -> None:
+            fired.append("SubroutineParameter")
+            self.generic_visit(node)
+
+        def visit_RangeDefinition(self, node: Any) -> None:
+            fired.append("RangeDefinition")
+            self.generic_visit(node)
+
+        def visit_SwitchCase(self, node: Any) -> None:
+            fired.append("SwitchCase")
+            self.generic_visit(node)
+
+        def visit_QuantumGateModifier(self, node: Any) -> None:
+            fired.append("QuantumGateModifier")
+            self.generic_visit(node)
+
+    result = semantic.analyze(source)
+    assert not result.has_errors
+    AuxiliaryVisitor().visit(result.program)
+    assert fired == [
+        "Annotation",
+        "SubroutineParameter",
+        "RangeDefinition",
+        "SwitchCase",
+        "QuantumGateModifier",
+    ]
